+++ Runtime/Waypoints.cs
@@ -10,48 +10,48 @@ namespace DumbTrollface.Waypoints
         /// </summary>
         [SerializeField]
         [Tooltip("Should the path be considered as closed (connection from the last to first waypoint) or open.")]
-        private bool closed = false;
+        private bool _closed = false;
 
         /// <summary>
-        /// List of all waypoints filled with some initial waypoints.
+        /// Array of all waypoints filled with some initial waypoints.
         /// </summary>
         [SerializeField]
-        private List<Vector3> waypoints = new()
+        private Vector3[] _waypoints = 
         {
-            new Vector3(-2f, 0f, -2f),
-            new Vector3( 2f, 0f, -2f),
-            new Vector3( 2f, 0f,  2f),
-            new Vector3(-2f, 0f,  2f)
+            new(-2f, 0f, -2f),
+            new( 2f, 0f, -2f),
+            new( 2f, 0f,  2f),
+            new(-2f, 0f,  2f)
         };
 
         /// <summary>
-        /// Return a readonly List of the waypoints.
+        /// Returns a copy of the waypoints.
         /// </summary>
-        public IReadOnlyList<Vector3> WaypointsList => waypoints;
+        public Vector3[] WaypointsCopy => (Vector3[])_waypoints.Clone();
 
         /// <summary>
         /// Returns the number of waypoints in this path
         /// </summary>
-        public int Count => waypoints.Count;
+        public int Count => _waypoints.Length;
 
         /// <summary>
         /// Returns the state of the closed setting
         /// </summary>
-        public bool Closed => closed;
+        public bool Closed => _closed;
 
         /// <summary>
         /// Returns the position of a waypoint in local space
         /// </summary>
         /// <param name="index">index of the waypoint</param>
         /// <returns>position in local space</returns>
-        public Vector3 GetLocalPoint(int index) => waypoints[index];
+        public Vector3 GetLocalPoint(int index) => _waypoints[index];
 
         /// <summary>
         /// Returns the position of a waypoint in world space
         /// </summary>
         /// <param name="index">index of the waypoint</param>
         /// <returns>position in world space</returns>
-        public Vector3 GetWorldPoint(int index) => transform.TransformPoint(waypoints[index]);
+        public Vector3 GetWorldPoint(int index) => transform.TransformPoint(_waypoints[index]);
     }
 
 }